backend/python_services/jwttoken_service/requirements.txt
fastapi[all]
sqlalchemy
asyncpg
sqlalchemy-utils
aiohttp
redis
pyjwt